requirements.txt
anyio==3.6.1
ariadne==0.15.1
arrow==1.2.2
asgiref==3.5.2
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cuid==0.3
Flask==2.2.0
Flask-Script==2.0.6
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-pubsub==2.13.6
googleapis-common-protos==1.56.4
graphql-core==3.2.1
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.22.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.20.4
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
Werkzeug==2.2.1
zipp==3.8.1
